refactor(services): terminal_service 를 모듈 4종으로 분할
867줄 → base 196 · adapters 433 · orca 194 · service 68 · __init__ 22.
전부 800 이내. **B5 대상 5개 전부 한도 안으로 들어왔다.**

정의 이동만 했고 본문은 한 글자도 바꾸지 않았다 (split_audit 으로 정의
24개 · 모듈레벨 할당 6개 전부 본문 일치, ruff format 이후 재실행도 0건).

## 다섯 번째 패치 형태 — 상수 조립

테스트가 MODULE = 'services.terminal_service' 를 두고 f"{MODULE}.xxx" 로
경로를 만든다. **타깃 문자열 grep 으로도 안 잡힌다** — 소스에 완성된 경로가
존재하지 않는다. 모듈 경로 상수를 먼저 찾는 것이 유일하게 확실한 스캔이다.

역설적으로 갱신은 한 줄이었다. 패치 타깃 4종이 전부 Orca 경로에 모여 있다:
  shutil             OrcaAdapter.is_available() 이 shutil.which 호출
  sys                _resolve_orca_command 가 sys.platform 읽음
  asyncio            _run_orca_json · _terminate_orca_process
  _write_exec_script OrcaAdapter 가 from .base import 로 가져옴 (모듈 지역)

앞 셋은 공유 모듈 객체 패치라 관대하고 네 번째는 모듈 지역이라 비관대다 —
그래서 경로가 orca 여야 한다. 넷 다 orca 가 갖는 이름이라 MODULE 을
'services.terminal_service.orca' 로 바꾸는 것만으로 충족됐다.

Red-Green: 갱신 전 19 failed / 4 passed → 갱신 후 23 passed.

## 배정 근거

_terminal_service 는 get_terminal_service 가 global 로 재바인딩하므로 service 에
함께 남겼다. TERMINAL_INFO 는 가변 dict 지만 사용처 전수가 읽기뿐이라 모듈을
갈라도 안전하며, api/terminal.py 와 테스트가 직접 import 하므로 재노출한다.
헬퍼 4종(_save_prompt_and_build_cmd·_build_full_command·_write_exec_script·
_run_osascript)은 base 에 모았다 — _write_exec_script 를 adapters 와 orca 가
공유하므로 공유 위치가 필요하고 나머지도 같은 성격이다.

게이트: ruff 0 · ruff format 0 · mypy 317파일 0 ·
pytest 1 failed, 1362 passed, 2 skipped (베이스라인 유지)
