+++ src/backend/services/terminal_service/__init__.py
@@ -1,867 +1,22 @@
-"""Terminal abstraction layer for multi-terminal execution.
+"""터미널 서비스 패키지.
 
-Provides a unified interface for executing commands across 8 terminal types:
-Warp, tmux, Terminal.app, iTerm2, Kitty, Alacritty, Ghostty, WezTerm.
+원래 단일 `services/terminal_service.py`(867줄)를 분할한 결과.
+소비자의 `from services.terminal_service import TERMINAL_INFO` 는 그대로 유효하다.
 
-WarpAdapter delegates to existing warp_service.py.
-TmuxAdapter delegates to existing tmux_service.py patterns.
-Other adapters implement terminal-specific execution via AppleScript or CLI.
+재노출은 **좁게** 한다 — 소비자가 실제로 요구하는 6종뿐이다.
+`shutil`·`sys`·`asyncio`·`_write_exec_script` 를 여기 두지 않는 것은 의도다:
+테스트의 패치는 그 이름들을 실제로 쓰는 `orca` 경로를 겨냥한다.
 """
 
-from __future__ import annotations
-
-import asyncio
-import json
-import logging
-import os
-import shlex
-import shutil
-import subprocess
-import sys
-import time
-from abc import ABC, abstractmethod
-from datetime import datetime
-from enum import Enum
-from pathlib import Path
-from typing import NamedTuple
-
-logger = logging.getLogger(__name__)
-
-IS_DOCKER = bool(os.getenv("CLAUDE_HOME"))
-
-
-class TerminalType(str, Enum):
-    """Supported terminal types."""
-
-    WARP = "warp"
-    TMUX = "tmux"
-    TERMINAL_APP = "terminal_app"
-    ITERM2 = "iterm2"
-    KITTY = "kitty"
-    ALACRITTY = "alacritty"
-    GHOSTTY = "ghostty"
-    WEZTERM = "wezterm"
-    CMUX = "cmux"
-    ORCA = "orca"
-
-
-TERMINAL_INFO: dict[TerminalType, dict[str, str]] = {
-    TerminalType.WARP: {"name": "Warp", "description": "AI-powered terminal"},
-    TerminalType.TMUX: {"name": "tmux", "description": "Terminal multiplexer"},
-    TerminalType.TERMINAL_APP: {
-        "name": "Terminal.app",
-        "description": "macOS built-in terminal",
-    },
-    TerminalType.ITERM2: {
-        "name": "iTerm2",
-        "description": "macOS terminal emulator",
-    },
-    TerminalType.KITTY: {"name": "Kitty", "description": "GPU-based terminal"},
-    TerminalType.ALACRITTY: {
-        "name": "Alacritty",
-        "description": "GPU-accelerated terminal",
-    },
-    TerminalType.GHOSTTY: {
-        "name": "Ghostty",
-        "description": "Fast, native terminal",
-    },
-    TerminalType.WEZTERM: {
-        "name": "WezTerm",
-        "description": "GPU-accelerated terminal by Wez Furlong",
-    },
-    TerminalType.CMUX: {
-        "name": "cmux",
-        "description": "AI-native terminal with workspaces",
-    },
-    TerminalType.ORCA: {
-        "name": "Orca",
-        "description": "AI-native workspace manager",
-    },
-}
-
-
-class TerminalAdapter(ABC):
-    """Abstract base class for terminal adapters."""
-
-    @abstractmethod
-    async def is_available(self) -> bool:
-        """Check whether this terminal is installed and usable."""
-
-    @abstractmethod
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        """Execute a command in this terminal.
-
-        Args:
-            project_path: Working directory for the command.
-            command: The prompt or command to execute.
-            title: Optional window/tab title.
-            branch_name: Git branch to create before execution.
-            image_paths: Image paths to pass via --image flags.
-
-        Returns:
-            dict with ``success``, ``terminal``, and ``message`` or ``error`` keys.
-        """
-
-
-# ---------------------------------------------------------------------------
-# Helper: build a claude CLI command string
-# ---------------------------------------------------------------------------
-
-
-def _save_prompt_and_build_cmd(
-    command: str,
-    image_paths: list[str] | None = None,
-) -> str:
-    """Save a prompt to a temp file and return a shell command string.
-
-    Mirrors the prompt-file pattern used in ``warp_service.py`` to avoid
-    shell-escaping issues with complex prompts.
-    """
-    prompt_dir = Path.home() / ".aos" / "prompts"
-    prompt_dir.mkdir(parents=True, exist_ok=True)
-
-    timestamp = datetime.now().strftime("%Y%m%d%H%M%S%f")
-    prompt_file = prompt_dir / f"aos-prompt-{timestamp}.txt"
-    prompt_file.write_text(command, encoding="utf-8")
-
-    base_cmd = "claude --dangerously-skip-permissions"
-    if image_paths:
-        image_flags = " ".join(f'--image "{p}"' for p in image_paths)
-        base_cmd = f"{base_cmd} {image_flags}"
-
-    return f'{base_cmd} "$(cat {prompt_file})"'
-
-
-def _build_full_command(
-    command: str,
-    branch_name: str | None = None,
-    image_paths: list[str] | None = None,
-) -> str:
-    """Build the full shell command including optional git checkout and claude CLI."""
-    claude_cmd = _save_prompt_and_build_cmd(command, image_paths)
-    if branch_name:
-        return f"git checkout -b {shlex.quote(branch_name)} && {claude_cmd}"
-    return claude_cmd
-
-
-def _write_exec_script(
-    project_path: str,
-    command: str,
-    branch_name: str | None = None,
-    image_paths: list[str] | None = None,
-) -> Path:
-    """Write a temp shell script for AppleScript-based terminals.
-
-    Complex commands with nested quotes break AppleScript string escaping.
-    Writing to a script file avoids all escaping issues.
-    """
-    full_command = _build_full_command(command, branch_name, image_paths)
-    script_dir = Path.home() / ".aos" / "scripts"
-    script_dir.mkdir(parents=True, exist_ok=True)
-
-    timestamp = datetime.now().strftime("%Y%m%d%H%M%S%f")
-    script_file = script_dir / f"aos-exec-{timestamp}.sh"
-    script_file.write_text(
-        f"#!/bin/bash\ncd {shlex.quote(project_path)} && {full_command}\n",
-        encoding="utf-8",
-    )
-    script_file.chmod(0o755)
-    return script_file
-
-
-# ---------------------------------------------------------------------------
-# Concrete adapters
-# ---------------------------------------------------------------------------
-
-
-class WarpAdapter(TerminalAdapter):
-    """Delegates to the existing WarpService."""
-
-    async def is_available(self) -> bool:
-        from services.warp_service import get_warp_service
-
-        return get_warp_service().is_warp_installed()
-
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        from services.warp_service import get_warp_service
-
-        warp = get_warp_service()
-        claude_cmd = warp.build_claude_command(task=command, image_paths=image_paths)
-        result = warp.open_with_command(
-            path=project_path,
-            command=claude_cmd,
-            title=title,
-            new_window=False,
-            branch_name=branch_name,
-        )
-        return {
-            "success": result.get("success", False),
-            "terminal": TerminalType.WARP.value,
-            "message": result.get("message"),
-            "error": result.get("error"),
-        }
-
-
-class TmuxAdapter(TerminalAdapter):
-    """Creates a tmux session and sends commands to it."""
-
-    async def is_available(self) -> bool:
-        return shutil.which("tmux") is not None
-
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        timestamp = int(time.time())
-        session_name = f"aos-{timestamp}"
-
-        try:
-            subprocess.run(
-                ["tmux", "new-session", "-d", "-s", session_name, "-c", project_path],
-                check=True,
-                capture_output=True,
-                text=True,
-            )
-        except subprocess.CalledProcessError as e:
-            logger.error("Failed to create tmux session: %s", e.stderr)
-            return {
-                "success": False,
-                "terminal": TerminalType.TMUX.value,
-                "error": f"Failed to create tmux session: {e.stderr}",
-            }
-
-        # Optional branch creation
-        if branch_name:
-            try:
-                subprocess.run(
-                    [
-                        "tmux",
-                        "send-keys",
-                        "-t",
-                        session_name,
-                        f"git checkout -b {shlex.quote(branch_name)}",
-                        "Enter",
-                    ],
-                    check=True,
-                    capture_output=True,
-                    text=True,
-                )
-            except subprocess.CalledProcessError as e:
-                logger.warning("Branch checkout send failed: %s", e.stderr)
-
-        # Build and send the claude command
-        claude_cmd = _save_prompt_and_build_cmd(command, image_paths)
-        try:
-            subprocess.run(
-                ["tmux", "send-keys", "-t", session_name, claude_cmd, "Enter"],
-                check=True,
-                capture_output=True,
-                text=True,
-            )
-        except subprocess.CalledProcessError as e:
-            logger.error("Failed to send command to tmux: %s", e.stderr)
-            return {
-                "success": False,
-                "terminal": TerminalType.TMUX.value,
-                "error": f"Failed to send command: {e.stderr}",
-            }
-
-        # Open a GUI terminal and attach to the tmux session so the user
-        # gets a visible window (tmux alone is detached/invisible).
-        # Prefer iTerm over Terminal.app when available.
-        attach_cmd = f"tmux attach -t {session_name}"
-        if Path("/Applications/iTerm.app").exists():
-            attach_script = (
-                'tell application "iTerm"\n'
-                "    activate\n"
-                "    create window with default profile\n"
-                "    tell current session of current window\n"
-                f'        write text "{attach_cmd}"\n'
-                "    end tell\n"
-                "end tell"
-            )
-        else:
-            attach_script = (
-                f'tell application "Terminal"\n    activate\n    do script "{attach_cmd}"\nend tell'
-            )
-        try:
-            subprocess.run(
-                ["osascript", "-e", attach_script],
-                check=True,
-                capture_output=True,
-                text=True,
-                timeout=10,
-            )
-        except (subprocess.CalledProcessError, subprocess.TimeoutExpired) as e:
-            logger.warning("Failed to auto-attach tmux session: %s", e)
-
-        return {
-            "success": True,
-            "terminal": TerminalType.TMUX.value,
-            "message": f"Started tmux session '{session_name}' at {project_path}",
-        }
-
-
-class TerminalAppAdapter(TerminalAdapter):
-    """macOS built-in Terminal.app via AppleScript."""
-
-    async def is_available(self) -> bool:
-        # Terminal.app is always available on macOS
-        return True
-
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        exec_script = _write_exec_script(project_path, command, branch_name, image_paths)
-        script = (
-            'tell application "Terminal"\n'
-            "    activate\n"
-            f'    do script "bash {exec_script}"\n'
-            "end tell"
-        )
-        return await _run_osascript(script, TerminalType.TERMINAL_APP)
-
-
-class ITermAdapter(TerminalAdapter):
-    """iTerm2 via AppleScript."""
-
-    async def is_available(self) -> bool:
-        return Path("/Applications/iTerm.app").exists()
-
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        exec_script = _write_exec_script(project_path, command, branch_name, image_paths)
-        script = (
-            'tell application "iTerm"\n'
-            "    activate\n"
-            "    create window with default profile\n"
-            "    tell current session of current window\n"
-            f'        write text "bash {exec_script}"\n'
-            "    end tell\n"
-            "end tell"
-        )
-        return await _run_osascript(script, TerminalType.ITERM2)
-
-
-class KittyAdapter(TerminalAdapter):
-    """Kitty terminal via CLI."""
-
-    async def is_available(self) -> bool:
-        return shutil.which("kitty") is not None
-
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        full_command = _build_full_command(command, branch_name, image_paths)
-        try:
-            subprocess.Popen(
-                [
-                    "kitty",
-                    "--single-instance",
-                    f"--directory={project_path}",
-                    "-e",
-                    "bash",
-                    "-c",
-                    full_command,
-                ],
-            )
-            return {
-                "success": True,
-                "terminal": TerminalType.KITTY.value,
-                "message": f"Opened Kitty at {project_path}",
-            }
-        except OSError as e:
-            logger.error("Failed to launch Kitty: %s", e)
-            return {
-                "success": False,
-                "terminal": TerminalType.KITTY.value,
-                "error": f"Failed to launch Kitty: {e}",
-            }
-
-
-class AlacrittyAdapter(TerminalAdapter):
-    """Alacritty terminal via CLI."""
-
-    async def is_available(self) -> bool:
-        return shutil.which("alacritty") is not None
-
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        full_command = _build_full_command(command, branch_name, image_paths)
-        try:
-            subprocess.Popen(
-                [
-                    "alacritty",
-                    "--working-directory",
-                    project_path,
-                    "-e",
-                    "bash",
-                    "-c",
-                    full_command,
-                ],
-            )
-            return {
-                "success": True,
-                "terminal": TerminalType.ALACRITTY.value,
-                "message": f"Opened Alacritty at {project_path}",
-            }
-        except OSError as e:
-            logger.error("Failed to launch Alacritty: %s", e)
-            return {
-                "success": False,
-                "terminal": TerminalType.ALACRITTY.value,
-                "error": f"Failed to launch Alacritty: {e}",
-            }
-
-
-class GhosttyAdapter(TerminalAdapter):
-    """Ghostty via AppleScript (System Events keystrokes).
-
-    Ghostty does not yet expose a rich AppleScript API, so we
-    activate the app and send keystrokes through System Events.
-    """
-
-    async def is_available(self) -> bool:
-        return Path("/Applications/Ghostty.app").exists()
-
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        exec_script = _write_exec_script(project_path, command, branch_name, image_paths)
-        script = (
-            'tell application "Ghostty"\n'
-            "    activate\n"
-            "end tell\n"
-            "delay 0.5\n"
-            'tell application "System Events"\n'
-            '    tell process "Ghostty"\n'
-            f'        keystroke "bash {exec_script}"\n'
-            "        key code 36\n"
-            "    end tell\n"
-            "end tell"
-        )
-        return await _run_osascript(script, TerminalType.GHOSTTY)
-
-
-class WezTermAdapter(TerminalAdapter):
-    """WezTerm via its CLI."""
-
-    async def is_available(self) -> bool:
-        return shutil.which("wezterm") is not None
-
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        full_command = _build_full_command(command, branch_name, image_paths)
-        try:
-            subprocess.Popen(
-                [
-                    "wezterm",
-                    "cli",
-                    "spawn",
-                    "--cwd",
-                    project_path,
-                    "--",
-                    "bash",
-                    "-c",
-                    full_command,
-                ],
-            )
-            return {
-                "success": True,
-                "terminal": TerminalType.WEZTERM.value,
-                "message": f"Opened WezTerm at {project_path}",
-            }
-        except OSError as e:
-            logger.error("Failed to launch WezTerm: %s", e)
-            return {
-                "success": False,
-                "terminal": TerminalType.WEZTERM.value,
-                "error": f"Failed to launch WezTerm: {e}",
-            }
-
-
-class CmuxAdapter(TerminalAdapter):
-    """cmux terminal via its CLI."""
-
-    async def is_available(self) -> bool:
-        return shutil.which("cmux") is not None or Path("/Applications/cmux.app").exists()
-
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        exec_script = _write_exec_script(project_path, command, branch_name, image_paths)
-
-        # cmux CLI IPC fails with Broken pipe from uvicorn subprocess.
-        # Workaround: open via LaunchServices, then type command via
-        # AppleScript keystroke (same approach as GhosttyAdapter).
-        try:
-            open_proc = await asyncio.create_subprocess_exec(
-                "/usr/bin/open",
-                "-a",
-                "cmux",
-                project_path,
-                stdout=asyncio.subprocess.DEVNULL,
-                stderr=asyncio.subprocess.PIPE,
-            )
-            _, open_err = await asyncio.wait_for(open_proc.communicate(), timeout=10)
-        except (TimeoutError, OSError) as e:
-            logger.error("Failed to launch cmux: %s", e)
-            return {
-                "success": False,
-                "terminal": TerminalType.CMUX.value,
-                "error": f"Failed to launch cmux: {e}",
-            }
-
-        if open_proc.returncode != 0:
-            err = open_err.decode().strip() if open_err else "unknown error"
-            logger.error("open -a cmux failed: %s", err)
-            return {
-                "success": False,
-                "terminal": TerminalType.CMUX.value,
-                "error": f"Failed to launch cmux: {err}",
-            }
-
-        # Type the command into the new workspace via System Events
-        script = (
-            'tell application "cmux"\n'
-            "    activate\n"
-            "end tell\n"
-            "delay 1.5\n"
-            'tell application "System Events"\n'
-            '    tell process "cmux"\n'
-            f'        keystroke "bash {exec_script}"\n'
-            "        key code 36\n"
-            "    end tell\n"
-            "end tell"
-        )
-        return await _run_osascript(script, TerminalType.CMUX)
-
-
-# ---------------------------------------------------------------------------
-# Orca CLI helpers
-# ---------------------------------------------------------------------------
-
-_ORCA_TIMEOUT_SECONDS = 20
-
-# Orca only resolves worktree selectors for repos it has registered. These
-# error codes mean "register the repo first", which is recoverable once.
-_ORCA_UNREGISTERED_ERROR_CODES = frozenset(
-    {"selector_not_found", "worktree_not_found", "repo_not_found"}
-)
-
-
-def _resolve_orca_command() -> str:
-    """Return the Orca CLI executable name for this platform.
-
-    On Linux a bare ``orca`` is the GNOME screen reader, not this app, so
-    the IDE ships its CLI as ``orca-ide``. ``ORCA_CLI_COMMAND`` overrides
-    both for non-standard installs. Resolved per call so environment and
-    platform changes are picked up without reimporting.
-    """
-    override = os.getenv("ORCA_CLI_COMMAND")
-    if override:
-        return override
-    return "orca" if sys.platform == "darwin" else "orca-ide"
-
-
-class _OrcaResult(NamedTuple):
-    """Outcome of a single ``orca ... --json`` invocation."""
-
-    ok: bool
-    code: str | None
-    error: str | None
-
-
-def _is_unregistered_worktree_error(result: _OrcaResult) -> bool:
-    """Return True when the failure means the repo is unknown to Orca."""
-    if result.code in _ORCA_UNREGISTERED_ERROR_CODES:
-        return True
-    text = (result.error or "").lower()
-    return "not found" in text and ("worktree" in text or "repo" in text)
-
-
-def _parse_orca_payload(stdout: str, stderr: str, returncode: int | None) -> _OrcaResult:
-    """Parse an Orca JSON envelope into an ``_OrcaResult``.
-
-    Orca emits ``{"ok": true, "result": {...}}`` on success and
-    ``{"ok": false, "error": {"code": ..., "message": ...}}`` on failure,
-    always on stdout, exiting non-zero for the latter.
-    """
-    if not stdout:
-        detail = stderr or "no output"
-        return _OrcaResult(False, None, f"orca exited with {returncode}: {detail}")
-
-    try:
-        payload = json.loads(stdout)
-    except json.JSONDecodeError:
-        logger.error("orca returned non-JSON output (exit %s)", returncode)
-        return _OrcaResult(False, None, f"failed to parse orca JSON response (exit {returncode})")
-
-    if payload.get("ok"):
-        return _OrcaResult(True, None, None)
-
-    error = payload.get("error")
-    if isinstance(error, dict):
-        code = error.get("code")
-        message = error.get("message") or code
-    else:
-        code = None
-        message = str(error) if error else None
-    return _OrcaResult(False, code, message or f"orca command failed (exit {returncode})")
-
-
-async def _terminate_orca_process(proc: asyncio.subprocess.Process) -> None:
-    """Kill and reap a child that outlived its timeout.
-
-    Without this the abandoned ``orca`` process keeps running and can
-    create the terminal *after* the API already reported failure.
-    """
-    try:
-        proc.kill()
-    except ProcessLookupError:
-        return  # already exited; nothing to reap
-    await proc.wait()
-
-
-async def _run_orca_json(args: list[str]) -> _OrcaResult:
-    """Run an ``orca`` CLI command with ``--json`` and parse its response.
-
-    ``stdin`` is detached: CLIs inheriting uvicorn's stdin have hung or
-    failed with broken pipes in this codebase before.
-    """
-    try:
-        proc = await asyncio.create_subprocess_exec(
-            *args,
-            stdin=asyncio.subprocess.DEVNULL,
-            stdout=asyncio.subprocess.PIPE,
-            stderr=asyncio.subprocess.PIPE,
-        )
-    except OSError as e:
-        logger.error("Failed to launch orca: %s", e)
-        return _OrcaResult(False, None, str(e))
-
-    try:
-        raw_out, raw_err = await asyncio.wait_for(proc.communicate(), timeout=_ORCA_TIMEOUT_SECONDS)
-    except TimeoutError:
-        logger.error("orca command timed out after %ss", _ORCA_TIMEOUT_SECONDS)
-        await _terminate_orca_process(proc)
-        return _OrcaResult(False, None, f"orca command timed out after {_ORCA_TIMEOUT_SECONDS}s")
-    except OSError as e:
-        logger.error("orca command failed while reading output: %s", e)
-        await _terminate_orca_process(proc)
-        return _OrcaResult(False, None, str(e))
-
-    stdout = raw_out.decode().strip() if raw_out else ""
-    stderr = raw_err.decode().strip() if raw_err else ""
-    return _parse_orca_payload(stdout, stderr, proc.returncode)
-
-
-class OrcaAdapter(TerminalAdapter):
-    """Orca workspace manager via its CLI.
-
-    Spawns a terminal session inside the existing checkout with
-    ``orca terminal create``; no new worktree is created.
-    """
-
-    async def is_available(self) -> bool:
-        return shutil.which(_resolve_orca_command()) is not None
-
-    async def execute(
-        self,
-        project_path: str,
-        command: str,
-        title: str | None = None,
-        branch_name: str | None = None,
-        image_paths: list[str] | None = None,
-    ) -> dict:
-        exec_script = _write_exec_script(project_path, command, branch_name, image_paths)
-
-        orca_cmd = _resolve_orca_command()
-        create_args = [
-            orca_cmd,
-            "terminal",
-            "create",
-            "--worktree",
-            f"path:{project_path}",
-            "--command",
-            f"bash {exec_script}",
-        ]
-        if title:
-            create_args += ["--title", title]
-        create_args.append("--json")
-
-        result = await _run_orca_json(create_args)
-
-        # Recover once from an unregistered repo by adding it, then retrying.
-        if not result.ok and _is_unregistered_worktree_error(result):
-            added = await _run_orca_json(
-                [orca_cmd, "repo", "add", "--path", project_path, "--json"]
-            )
-            if added.ok:
-                result = await _run_orca_json(create_args)
-            else:
-                result = added
-
-        if result.ok:
-            return {
-                "success": True,
-                "terminal": TerminalType.ORCA.value,
-                "message": f"Created Orca terminal at {project_path}",
-            }
-
-        logger.error("orca terminal create failed: %s", result.error)
-        return {
-            "success": False,
-            "terminal": TerminalType.ORCA.value,
-            "error": f"Failed to create Orca terminal: {result.error}",
-        }
-
-
-# ---------------------------------------------------------------------------
-# Shared helper for AppleScript-based adapters
-# ---------------------------------------------------------------------------
-
-
-async def _run_osascript(script: str, terminal_type: TerminalType) -> dict:
-    """Run an AppleScript and return a standardised result dict."""
-    try:
-        subprocess.run(
-            ["osascript", "-e", script],
-            check=True,
-            capture_output=True,
-            text=True,
-            timeout=10,
-        )
-        name = TERMINAL_INFO[terminal_type]["name"]
-        return {
-            "success": True,
-            "terminal": terminal_type.value,
-            "message": f"Opened {name}",
-        }
-    except subprocess.TimeoutExpired:
-        return {
-            "success": False,
-            "terminal": terminal_type.value,
-            "error": "AppleScript timed out",
-        }
-    except subprocess.CalledProcessError as e:
-        logger.error("osascript failed for %s: %s", terminal_type.value, e)
-        return {
-            "success": False,
-            "terminal": terminal_type.value,
-            "error": f"AppleScript execution failed: {e}",
-        }
-
-
-# ---------------------------------------------------------------------------
-# Service
-# ---------------------------------------------------------------------------
-
-
-class TerminalService:
-    """Facade that routes execution to the correct terminal adapter."""
-
-    def __init__(self) -> None:
-        self._adapters: dict[TerminalType, TerminalAdapter] = {
-            TerminalType.WARP: WarpAdapter(),
-            TerminalType.TMUX: TmuxAdapter(),
-            TerminalType.TERMINAL_APP: TerminalAppAdapter(),
-            TerminalType.ITERM2: ITermAdapter(),
-            TerminalType.KITTY: KittyAdapter(),
-            TerminalType.ALACRITTY: AlacrittyAdapter(),
-            TerminalType.GHOSTTY: GhosttyAdapter(),
-            TerminalType.WEZTERM: WezTermAdapter(),
-            TerminalType.CMUX: CmuxAdapter(),
-            TerminalType.ORCA: OrcaAdapter(),
-        }
-
-    def get_adapter(self, terminal: TerminalType) -> TerminalAdapter:
-        """Return the adapter for the given terminal type."""
-        return self._adapters[terminal]
-
-    async def detect_available(self) -> list[dict]:
-        """Probe every adapter and return availability info."""
-        results: list[dict] = []
-        for t, adapter in self._adapters.items():
-            info = TERMINAL_INFO[t]
-            available = await adapter.is_available()
-            results.append(
-                {
-                    "type": t.value,
-                    "name": info["name"],
-                    "description": info["description"],
-                    "available": available,
-                }
-            )
-        return results
-
-
-# ---------------------------------------------------------------------------
-# Singleton
-# ---------------------------------------------------------------------------
-
-_terminal_service: TerminalService | None = None
-
-
-def get_terminal_service() -> TerminalService:
-    """Return the TerminalService singleton."""
-    global _terminal_service
-    if _terminal_service is None:
-        _terminal_service = TerminalService()
-    return _terminal_service
+from .base import TERMINAL_INFO, TerminalType
+from .orca import OrcaAdapter, _resolve_orca_command
+from .service import TerminalService, get_terminal_service
+
+__all__ = [
+    "TERMINAL_INFO",
+    "OrcaAdapter",
+    "TerminalService",
+    "TerminalType",
+    "_resolve_orca_command",
+    "get_terminal_service",
+]
